refactor: 코딩 컨벤션 및 패키지 변경
