+++ src/main/java/lotto/util/LottoNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.util;
 
 import java.util.List;
