+++ src/main/java/lotto/util/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.util;
 
 import java.util.*;
 
@@ -7,10 +7,10 @@ public class RandomNumberGenerator implements LottoNumberGenerator {
     public static final int RANGE = 45;
 
     public List<Integer> getNumbers() {
-        Set<Integer> numbers = new HashSet<>();
+        Set<Integer> numbers = new TreeSet<>();
         Random random = new Random();
         while (numbers.size() < 6) {
-            numbers.add(random.nextInt(RANGE));
+            numbers.add(random.nextInt(RANGE) + 1);
         }
         return new ArrayList<>(numbers);
     }
