+++ src/test/java/lotto/util/RandomNumberTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.util;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
